proposing a new organization
